pipelines/vlm/requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
pillow>=10.0.0
requests>=2.31.0
tqdm>=4.66.0
jinja2>=3.1.0
